tests: Move ownership update coverage
Batch source refresh tests currently cover stale-source repair,
selected-line remapping, and ownership-update preparation.

That mixes the new ownership-update module's contract into a source-refresh
test file, so the test tree no longer mirrors the runtime owner.

This commit moves the prepared-update and command-helper coverage into
test_ownership_update and retargets monkeypatches and import expectations
to batch.ownership_update.

An architecture guard will next pin the ownership_update boundary and
reject stale source_refresh imports.
