+++ tests/batch/test_source_refresh.py
@@ -1,25 +1,10 @@
-"""Tests for batch source refresh helpers.
-
-These tests verify the centralized stale-source repair logic
-and ensure include/discard commands use the same helpers.
-"""
+"""Tests for batch source refresh helpers."""
 
 from __future__ import annotations
 
-
-import inspect
-import git_stage_batch.batch.source_refresh as source_refresh_module
-from git_stage_batch.commands.selection import (
-    selected_change_batch_discarding,
-    selected_change_batch_staging,
-)
-
 from git_stage_batch.batch.source_refresh import (
     RefreshedBatchSelection,
-    PreparedBatchUpdate,
-    acquire_batch_ownership_update_for_selection,
     ensure_batch_source_current_for_selection,
-    prepare_batch_ownership_update_for_selection,
 )
 from git_stage_batch.batch.selected_line_source_refresh import (
     refresh_selected_lines_against_source_lines,
@@ -62,21 +47,6 @@ def test_refreshed_batch_selection_dataclass():
     assert refresh.ownership is None
     assert refresh.selected_lines == []
     assert refresh.source_was_advanced is False
-
-
-def test_prepared_batch_update_dataclass():
-    """Test PreparedBatchUpdate dataclass construction."""
-    ownership = BatchOwnership.from_presence_lines(["1-3"], [])
-
-    update = PreparedBatchUpdate(
-        batch_source_commit="def456",
-        ownership_before=None,
-        ownership_after=ownership
-    )
-
-    assert update.batch_source_commit == "def456"
-    assert update.ownership_before is None
-    assert update.ownership_after == ownership
 
 
 def test_ensure_batch_source_current_non_stale_source():
@@ -130,175 +100,6 @@ def test_ensure_batch_source_current_first_time_stale():
     assert result.ownership is None
     assert result.selected_lines[0].source_line == 1
     assert result.source_was_advanced is False
-
-
-def test_prepare_batch_ownership_update_first_time_stale_blank_context():
-    """First-time stale selections re-annotate blank context before translation."""
-    lines = [
-        LineEntry(
-            id=None, kind=' ', old_line_number=1, new_line_number=1,
-            text_bytes=b"", text="", source_line=None
-        ),
-        LineEntry(
-            id=1, kind='+', old_line_number=None, new_line_number=2,
-            text_bytes=b"new line", text="new line", source_line=None
-        ),
-    ]
-
-    result = prepare_batch_ownership_update_for_selection(
-        batch_name="test-batch",
-        file_path="test.py",
-        current_batch_source_commit=None,
-        existing_ownership=None,
-        selected_lines=lines
-    )
-
-    assert result.batch_source_commit is None
-    assert result.ownership_before is None
-    assert result.ownership_after.presence_claims[0].source_lines == ["1-2"]
-
-
-def test_prepare_batch_ownership_update_first_time_deletion_anchor():
-    """First-time deletion-only selections keep their source anchor."""
-    lines = [
-        LineEntry(
-            id=1, kind='-', old_line_number=2, new_line_number=None,
-            text_bytes=b"old line", text="old line", source_line=None
-        ),
-    ]
-
-    result = prepare_batch_ownership_update_for_selection(
-        batch_name="test-batch",
-        file_path="test.py",
-        current_batch_source_commit=None,
-        existing_ownership=None,
-        selected_lines=lines
-    )
-
-    assert result.batch_source_commit is None
-    assert result.ownership_before is None
-    assert result.ownership_after.deletions[0].anchor_line == 1
-
-
-def test_prepare_batch_ownership_update_first_time():
-    """Test prepare_batch_ownership_update_for_selection for first-time add."""
-    # Lines with valid source_line (first add to batch)
-    lines = [
-        LineEntry(
-            id=1, kind='+', old_line_number=None, new_line_number=1,
-            text_bytes=b"line1\n", text="line1\n", source_line=1
-        ),
-        LineEntry(
-            id=2, kind='+', old_line_number=None, new_line_number=2,
-            text_bytes=b"line2\n", text="line2\n", source_line=2
-        ),
-    ]
-
-    result = prepare_batch_ownership_update_for_selection(
-        batch_name="test-batch",
-        file_path="test.py",
-        current_batch_source_commit=None,
-        existing_ownership=None,
-        selected_lines=lines
-    )
-
-    assert result.batch_source_commit is None
-    assert result.ownership_before is None
-    assert result.ownership_after is not None
-    assert result.ownership_after.presence_claims[0].source_lines == ["1-2"]
-
-
-def test_prepare_batch_ownership_update_with_existing():
-    """Test prepare_batch_ownership_update_for_selection merging with existing ownership."""
-    # Existing ownership claims lines 1-2
-    existing = BatchOwnership.from_presence_lines(["1-2"], [])
-
-    # New lines claim lines 3-4
-    lines = [
-        LineEntry(
-            id=3, kind='+', old_line_number=None, new_line_number=3,
-            text_bytes=b"line3\n", text="line3\n", source_line=3
-        ),
-        LineEntry(
-            id=4, kind='+', old_line_number=None, new_line_number=4,
-            text_bytes=b"line4\n", text="line4\n", source_line=4
-        ),
-    ]
-
-    result = prepare_batch_ownership_update_for_selection(
-        batch_name="test-batch",
-        file_path="test.py",
-        current_batch_source_commit="source123",
-        existing_ownership=existing,
-        selected_lines=lines
-    )
-
-    assert result.ownership_before == existing
-    assert result.ownership_after is not None
-    # Should merge 1-2 with 3-4
-    assert "1-4" in ",".join(result.ownership_after.presence_claims[0].source_lines)
-
-
-def test_acquire_batch_ownership_update_uses_metadata_acquisition(monkeypatch):
-    """Prepared updates can borrow ownership from metadata while open."""
-    existing = BatchOwnership.from_presence_lines(["1"], [])
-    entered = False
-    exited = False
-
-    class OwnershipContext:
-        def __enter__(self):
-            nonlocal entered
-            entered = True
-            return existing
-
-        def __exit__(self, exc_type, exc, traceback):
-            nonlocal exited
-            exited = True
-
-    def acquire_for_metadata_dict(metadata):
-        assert metadata == {"batch_source_commit": "source123"}
-        return OwnershipContext()
-
-    if hasattr(source_refresh_module, "acquire_ownership_for_metadata_dict"):
-        monkeypatch.setattr(
-            source_refresh_module,
-            "acquire_ownership_for_metadata_dict",
-            acquire_for_metadata_dict,
-        )
-    else:
-        def acquire_for_metadata_dict_method(cls, metadata):
-            return acquire_for_metadata_dict(metadata)
-
-        monkeypatch.setattr(
-            BatchOwnership,
-            "acquire_for_metadata_dict",
-            classmethod(acquire_for_metadata_dict_method),
-        )
-    lines = [
-        LineEntry(
-            id=2,
-            kind="+",
-            old_line_number=None,
-            new_line_number=2,
-            text_bytes=b"line2\n",
-            text="line2\n",
-            source_line=2,
-        ),
-    ]
-
-    with acquire_batch_ownership_update_for_selection(
-        batch_name="test-batch",
-        file_path="test.py",
-        file_metadata={"batch_source_commit": "source123"},
-        selected_lines=lines,
-    ) as result:
-        assert entered is True
-        assert exited is False
-        assert result.batch_source_commit == "source123"
-        assert result.ownership_before is existing
-        assert result.ownership_after.presence_line_set() == {1, 2}
-
-    assert exited is True
 
 
 def test_refresh_selected_lines_uses_synthesized_working_line_provenance():
@@ -364,31 +165,3 @@ def test_refresh_selected_lines_accepts_non_list_line_sequences(line_sequence):
     )
 
     assert refreshed[0].source_line == 3
-
-
-def test_both_commands_use_same_helper_interface():
-    """Test that both include and discard use acquired update preparation.
-
-    This test verifies the refactoring succeeded: both command paths now use
-    the same centralized helper instead of duplicated inline logic.
-    """
-
-    # Read the source code of include and discard modules
-
-    include_source = inspect.getsource(selected_change_batch_staging)
-    discard_source = inspect.getsource(selected_change_batch_discarding)
-
-    # Include and selected-change discard batching should import from batch.source_refresh.
-    assert "from ...batch.source_refresh import acquire_batch_ownership_update_for_selection" in include_source
-    assert "from ...batch.source_refresh import acquire_batch_ownership_update_for_selection" in discard_source
-
-    # Both paths should call acquire_batch_ownership_update_for_selection.
-    assert "acquire_batch_ownership_update_for_selection(" in include_source
-    assert "acquire_batch_ownership_update_for_selection(" in discard_source
-    assert "prepare_batch_ownership_update_for_selection(" not in include_source
-    assert "prepare_batch_ownership_update_for_selection(" not in discard_source
-
-    # Neither should have the old inline implementation pattern
-    # (checking they don't manually handle stale source advancement)
-    assert "ensure_batch_source_current_and_reannotate" not in include_source
-    assert "ensure_batch_source_current_and_reannotate" not in discard_source
